+++ frontend/src/apis/api/user.ts
@@ -1,5 +1,4 @@
-import { AxiosResponse } from 'axios';
-import { authInstance, defaultInstance } from '../utils/instance';
+import { authInstance } from '../utils/instance';
 
 export const getUserInfo = async () => {
   try {
@@ -8,19 +7,4 @@ export const getUserInfo = async () => {
   } catch (err) {
     console.log(err);
   }
-};
-
-export const userLogin = async (data: TLoginForm) => {
-  try {
-    const res: AxiosResponse = await defaultInstance.post('/auth/login', data);
-    return res;
-  } catch (err) {
-    console.log(err);
-    return err;
-  }
-};
-
-type TLoginForm = {
-  email: string;
-  password: string;
 };
